refactor(frontend): reorganize source layout for maintainability
