+++ frontend/src/components/ViewHost.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useMemo, useRef } from "react";
-import * as statusTracker from "../legacy/js/status_tracker";
-import { VIEW_LOADERS } from "../viewLoaders";
-import { VIEW_TEMPLATES } from "../viewTemplates";
-import type { AppRoute, RouteName, ViewLoaderModule } from "../types";
+import { VIEW_LOADERS } from "../app/viewLoaders";
+import { VIEW_TEMPLATES } from "../app/viewTemplates";
+import type { AppRoute, RouteName, ViewLoaderModule } from "../app/routeTypes";
+import * as statusTracker from "../services/statusTracker";
 
 function getTemplateHtml(routeName: RouteName): string {
   return VIEW_TEMPLATES[routeName] || VIEW_TEMPLATES.library;
