+++ frontend/src/views/library.ts
@@ -1,7 +1,7 @@
-import { apiDelete, apiGet, apiPost } from '../api';
-import { renderMediaCard } from '../components/media_card';
-import { notify } from '../notify';
-import * as statusTracker from '../status_tracker';
+import { renderMediaCard } from '../ui/mediaCard';
+import { apiDelete, apiGet, apiPost } from '../services/api';
+import { notify } from '../services/notify';
+import * as statusTracker from '../services/statusTracker';
 
 function normalizeLibraryItem(item) {
   return {
